saluki_api

Module extract

Modules

connect_info

Extractor for getting connection information from a client.

path

Extractor that will get captures from the URL and parse them using [serde].

rejection

Rejection response types.

Structs

ConnectInfo

Extractor for getting connection information produced by a Connected.

DefaultBodyLimit

Layer for configuring the default request body limit.

Extension

Extractor and response for extensions.

Json

JSON Extractor / Response.

NestedPath

Access the path the matched the route is nested at.

OriginalUri

Extractor that gets the original request URI regardless of nesting.

Path

Extractor that will get captures from the URL and parse them using serde.

Query

Extractor that deserializes query strings into some type.

RawForm

Extractor that extracts raw form requests.

RawPathParams

Extractor that will get captures from the URL without deserializing them.

RawQuery

Extractor that extracts the raw query string, without parsing it.

State

Extractor for state.

Traits

FromRef

Used to do reference-to-value conversions thus not consuming the input value.

FromRequest

Types that can be created from requests.

FromRequestParts

Types that can be created from request parts.

OptionalFromRequest

Customize the behavior of Option<Self> as a FromRequest extractor.

OptionalFromRequestParts

Customize the behavior of Option<Self> as a FromRequestParts extractor.

Type Aliases

Request

Type alias for http::Request whose body type defaults to Body, the most common body type used with axum.