saluki_api::extract

Module path

Extractor that will get captures from the URL and parse them using serde.

Structs

FailedToDeserializePathParams

Rejection type for Path if the captured routes params couldn’t be deserialized into the expected type.

InvalidUtf8InPathParam

Rejection used by RawPathParams if a parameter contained text that, once percent decoded, wasn’t valid UTF-8.

Path

Extractor that will get captures from the URL and parse them using serde.

RawPathParams

Extractor that will get captures from the URL without deserializing them.

RawPathParamsIter

An iterator over raw path parameters.

Enums

ErrorKind

The kinds of errors that can happen we deserializing into a Path.