saluki_api::extract

Module rejection

Rejection response types.

Structs

FailedToDeserializeForm

Rejection type used if the Form extractor is unable to deserialize the form into the target type.

FailedToDeserializeFormBody

Rejection type used if the Form extractor is unable to deserialize the form body into the target type.

FailedToDeserializePathParams

Rejection type for Path if the captured routes params couldn’t be deserialized into the expected type.

FailedToDeserializeQueryString

Rejection type used if the Query extractor is unable to deserialize the query string into the target type.

InvalidFormContentType

Rejection type for Form or RawForm used if the Content-Type header is missing or its value is not application/x-www-form-urlencoded.

InvalidUtf8

Rejection type used when buffering the request into a String if the body doesn’t contain valid UTF-8.

InvalidUtf8InPathParam

Rejection used by RawPathParams if a parameter contained text that, once percent decoded, wasn’t valid UTF-8.

JsonDataError

Rejection type for Json.

JsonSyntaxError

Rejection type for Json.

LengthLimitError

Encountered some other error when buffering the body.

MissingExtension

Rejection type for Extension if an expected request extension was not found.

MissingJsonContentType

Rejection type for Json used if the Content-Type header is missing.

MissingPathParams

Rejection type used if axum’s internal representation of path parameters is missing. This is commonly caused by extracting Request<_>. Path must be extracted first.

NestedPathRejection

Rejection type for NestedPath.

UnknownBodyError

Encountered an unknown error when buffering the body.

Enums

BytesRejection

Rejection used for Bytes.

ExtensionRejection

Rejection used for Extension.

FailedToBufferBody

Rejection type for extractors that buffer the request body. Used if the request body cannot be buffered due to an error.

FormRejection

Rejection used for Form.

JsonRejection

Rejection used for Json.

PathRejection

Rejection used for Path.

QueryRejection

Rejection used for Query.

RawFormRejection

Rejection used for RawForm.

RawPathParamsRejection

Rejection used for RawPathParams.

StringRejection

Rejection used for String.